React package built by SalesforceSDKManager
