+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/app/SalesforceSDKManager.java
@@ -48,6 +48,11 @@ import android.util.Log;
 import android.webkit.CookieManager;
 import android.webkit.CookieSyncManager;
 
+import com.facebook.react.ReactPackage;
+import com.facebook.react.bridge.JavaScriptModule;
+import com.facebook.react.bridge.NativeModule;
+import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.uimanager.ViewManager;
 import com.salesforce.androidsdk.accounts.UserAccount;
 import com.salesforce.androidsdk.accounts.UserAccountManager;
 import com.salesforce.androidsdk.auth.AuthenticatorService;
@@ -59,6 +64,8 @@ import com.salesforce.androidsdk.config.BootConfig;
 import com.salesforce.androidsdk.config.LoginServerManager;
 import com.salesforce.androidsdk.push.PushMessaging;
 import com.salesforce.androidsdk.push.PushNotificationInterface;
+import com.salesforce.androidsdk.reactnative.SalesforceNetReactBridge;
+import com.salesforce.androidsdk.reactnative.SalesforceOauthReactBridge;
 import com.salesforce.androidsdk.rest.ClientManager;
 import com.salesforce.androidsdk.rest.ClientManager.LoginOptions;
 import com.salesforce.androidsdk.security.Encryptor;
@@ -73,6 +80,8 @@ import com.salesforce.androidsdk.util.EventsObservable;
 import com.salesforce.androidsdk.util.EventsObservable.EventType;
 
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -1123,5 +1132,37 @@ public class SalesforceSDKManager {
 	        CookieSyncManager.createInstance(context);
 	        CookieSyncManager.getInstance().sync();
 		}
+    }
+
+    /**
+     * Call this method when setting up ReactInstanceManager
+     * @return ReactPackage for this application
+     */
+    public ReactPackage getReactPackage() {
+        if (appType!= AppType.ReactNative) {
+            return null;
+        }
+        return new ReactPackage() {
+            @Override
+            public List<NativeModule> createNativeModules(
+                    ReactApplicationContext reactContext) {
+                List<NativeModule> modules = new ArrayList<>();
+
+                modules.add(new SalesforceOauthReactBridge(reactContext));
+                modules.add(new SalesforceNetReactBridge(reactContext));
+
+                return modules;
+            }
+
+            @Override
+            public List<Class<? extends JavaScriptModule>> createJSModules() {
+                return Collections.emptyList();
+            }
+
+            @Override
+            public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
+                return Collections.emptyList();
+            }
+        };
     }
 }
